Avoid division by zero in metrics calculation.
Sometimes a selected tag is not present in the covered file (e.g. there are no branches). In this scenario the metrics should evaluate the passed/total ratio to one, not NaN.

The reason of 1 and not 0 is so that when there are no, let's say branches, then we indicate that they were all covered and don't show a 0% branch coverage error.